refactor: pois and loadLeaflet into own file
